Refactor map utility functions to separate file
Move utility functions to a separate file and update imports and tests accordingly.

* Add `scripts/map-utils.js` to define and export `getColor` and `getWeight` functions.
* Modify `scripts/map.js` to import `getColor` and `getWeight` from `./map-utils` and remove their definitions.
* Rename `tests/map.test.js` to `tests/map-utils.test.js` and update imports to use `../scripts/map-utils`.

---

For more details, open the [Copilot Workspace session](https://copilot-workspace.githubnext.com/statnmap/gpx-traces-website/tree/add-unit-tests?shareId=XXXX-XXXX-XXXX-XXXX).
